+++ api/src/measurement.rs
@@ -1,14 +1,15 @@
-pub mod composable;
-pub mod convertible;
-pub mod field_eq;
-pub mod invert;
-pub mod is_compatible_with;
-pub mod ops;
-pub mod partial_eq;
-pub mod partial_ord;
-pub mod reducible;
-pub mod to_reduced;
-pub mod ucum_unit;
+mod composable;
+mod convertible;
+mod display;
+mod field_eq;
+mod invert;
+mod is_compatible_with;
+mod ops;
+mod partial_eq;
+mod partial_ord;
+mod reducible;
+mod to_reduced;
+mod ucum_unit;
 
 use crate::error::Error;
 use crate::reducible::Reducible;
@@ -43,6 +44,10 @@ pub struct Measurement {
 
 impl Measurement {
     /// Creates a new `Measurement` by parsing `expression` into a `Unit`.
+    ///
+    /// # Errors
+    ///
+    /// Returns an `Error` if `expression` isn't one that represents a valid `Unit`.
     ///
     #[inline]
     pub fn new(value: f64, expression: &str) -> Result<Self, Error> {
@@ -167,13 +172,16 @@ mod tests {
         }
 
         #[test]
-        fn validate_deserialize_json_errors() {
-            let expected_json = r#"{"value":2.0,"unit":""}"#;
+        fn validate_deserialize_json_errors_bad_value_type() {
+            let expected_json = r#"{"value":"adsf","unit":"m"}"#;
             let measurement: Result<Measurement, serde_json::Error> =
                 serde_json::from_str(expected_json);
             assert!(measurement.is_err());
+        }
 
-            let expected_json = r#"{"value":"adsf","unit":"m"}"#;
+        #[test]
+        fn validate_deserialize_json_errors_empty_unit() {
+            let expected_json = r#"{"value":2.0,"unit":""}"#;
             let measurement: Result<Measurement, serde_json::Error> =
                 serde_json::from_str(expected_json);
             assert!(measurement.is_err());