+++ api/CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Measurement` had a submodule for implementing `Display`, but it wasn't being
   `use`d; could call this a "fix", but technically it's adding this impl.
 
+### Removed
+
+- Both `measurement` and `unit` had all submodules as `pub` that just implement
+  traits; this is unnecessary.
+
 ## [0.12.0] - 2019-11-07
 
 ### Added