+++ api/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unrelesaed]
+
+### Added
+
+- `Composition` now exposes `const fn`s for constructing 1-`Dimension`al
+  objects.
+
 ## [0.13.0] - 2020-03-20
 
 ### Added