+++ api/CHANGELOG.md
@@ -5,11 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unrelesaed]
+## [0.14.0] - 2020-05-08
 
 ### Added
 
-- `Composition` now exposes `const fn`s for constructing 1-`Dimension`al
+- `Composition` now exposes `const fn`s for constructing one-`Dimension`al
   objects.
 - `Composition` now derives `Eq`.
 - `Error` now derives `serde::Serialize`.